+++ src/Reactor1D.hpp
@@ -188,22 +188,22 @@ namespace CVI
 
 		// Mole fractions [-]
 		X_.resize(np_);
-		for (int i = 0; i < np_; i++)
+		for (unsigned int i = 0; i < np_; i++)
 			X_[i].resize(nc_);
 
 		// Mass fractions [-]
 		Y_.resize(np_);
-		for (int i = 0; i < np_; i++)
+		for (unsigned int i = 0; i < np_; i++)
 			Y_[i].resize(nc_);
 
 		// Formation rates in gas pahse [kg/m3/s]
 		omega_homogeneous_from_homogeneous_.resize(np_);
-		for (int i = 0; i < np_; i++)
+		for (unsigned int i = 0; i < np_; i++)
 			omega_homogeneous_from_homogeneous_[i].resize(nc_);
 
 		// Heterogeneous formation rates [kg/m3/s]
 		omega_homogeneous_from_heterogeneous_.resize(np_);
-		for (int i = 0; i < np_; i++)
+		for (unsigned int i = 0; i < np_; i++)
 			omega_homogeneous_from_heterogeneous_[i].resize(nc_);
 
 		// Deposition rate [kg/m3/s]
@@ -217,12 +217,12 @@ namespace CVI
 
 		// Effective diffusion coefficiennts [m2/s]
 		gamma_star_.resize(np_);
-		for (int i = 0; i < np_; i++)
+		for (unsigned int i = 0; i < np_; i++)
 			gamma_star_[i].resize(nc_);
 
 		// Diffusion fluxes [???]
 		j_star_.resize(np_ - 1);
-		for (int i = 0; i < np_ - 1; i++)
+		for (unsigned int i = 0; i < np_ - 1; i++)
 			j_star_[i].resize(nc_);
 
 		// Correction diffusion flux [???]
@@ -230,27 +230,27 @@ namespace CVI
 
 		// Spatial derivatives: mole fractions [1/m]
 		dX_over_dx_.resize(np_);
-		for (int i = 0; i < np_; i++)
+		for (unsigned int i = 0; i < np_; i++)
 			dX_over_dx_[i].resize(nc_);
 
 		// Spatial derivatives: mass fractions [1/m]
 		dY_over_dx_.resize(np_);
-		for (int i = 0; i < np_; i++)
+		for (unsigned int i = 0; i < np_; i++)
 			dY_over_dx_[i].resize(nc_);
 
 		// Second order spatial derivatives: mass fractions [1/m2]
 		d2Y_over_dx2_.resize(np_);
-		for (int i = 0; i < np_; i++)
+		for (unsigned int i = 0; i < np_; i++)
 			d2Y_over_dx2_[i].resize(nc_);
 
 		// Time derivatives: mass fractions [1/s]
 		dY_over_dt_.resize(np_);
-		for (int i = 0; i < np_; i++)
+		for (unsigned int i = 0; i < np_; i++)
 			dY_over_dt_[i].resize(nc_);
 
 		// Spatial derivatives: mass diffusion coefficients [m2/s/m]
 		dgamma_star_over_dx_.resize(np_);
-		for (int i = 0; i < np_; i++)
+		for (unsigned int i = 0; i < np_; i++)
 			dgamma_star_over_dx_[i].resize(nc_);
 
 		// Spatial derivatives: density of gaseous phase
@@ -260,11 +260,11 @@ namespace CVI
 		depsilon_over_dt_.resize(np_);
 
 		// Reset to zero
-		for (int i = 0; i < np_-1; i++)
+		for (unsigned int i = 0; i < np_-1; i++)
 			j_star_[i].setZero();
-		for (int i = 0; i < np_; i++)
+		for (unsigned int i = 0; i < np_; i++)
 			omega_homogeneous_from_homogeneous_[i].setZero();
-		for (int i = 0; i < np_; i++)
+		for (unsigned int i = 0; i < np_; i++)
 			omega_homogeneous_from_heterogeneous_[i].setZero();
 		omega_deposition_per_unit_volume_.setZero();
 		omega_deposition_per_unit_area_.setZero();
@@ -275,24 +275,24 @@ namespace CVI
 		{
 			// Surface fractions [-]
 			Z_.resize(np_);
-			for (int i = 0; i < np_; i++)
+			for (unsigned int i = 0; i < np_; i++)
 				Z_[i].resize(surf_nc_);
 
 			// Time derivatives: surface species fractions [1/s]
 			dZ_over_dt_.resize(np_);
-			for (int i = 0; i < np_; i++)
+			for (unsigned int i = 0; i < np_; i++)
 				dZ_over_dt_[i].resize(surf_nc_);
 
 			Gamma_.resize(np_);
-			for (int i = 0; i < np_; i++)
+			for (unsigned int i = 0; i < np_; i++)
 				Gamma_[i].resize(surf_np_);
 
 			GammaFromEqn_.resize(np_);
-			for (int i = 0; i < np_; i++)
+			for (unsigned int i = 0; i < np_; i++)
 				GammaFromEqn_[i].resize(surf_np_);
 
 			dGamma_over_dt_.resize(np_);
-			for (int i = 0; i < np_; i++)
+			for (unsigned int i = 0; i < np_; i++)
 			{
 				dGamma_over_dt_[i].resize(surf_np_);
 				dGamma_over_dt_[i].setZero();
@@ -300,7 +300,7 @@ namespace CVI
 
 			// Heterogeneous formation rates for surface species [kg/m2/s]
 			omega_heterogeneous_from_heterogeneous_.resize(np_);
-			for (int i = 0; i < np_; i++)
+			for (unsigned int i = 0; i < np_; i++)
 				omega_heterogeneous_from_heterogeneous_[i].resize(surf_nc_);
 
 			// Auxiliary vectors
@@ -310,7 +310,7 @@ namespace CVI
 			eigen_gamma_.resize(surf_np_);
 
 			// Reset to zero
-			for (int i = 0; i < np_; i++)
+			for (unsigned int i = 0; i < np_; i++)
 				omega_heterogeneous_from_heterogeneous_[i].setZero();
 		}
 	}
@@ -469,15 +469,15 @@ namespace CVI
 
 	void Reactor1D::SetInitialConditions(const double T_gas, const double P_gas, const Eigen::VectorXd& omega_gas, const Eigen::VectorXd& Gamma0, const Eigen::VectorXd& Z0)
 	{
-		for (int i = 0; i < np_; i++)
+		for (unsigned int i = 0; i < np_; i++)
 			for (unsigned int j = 0; j < nc_; j++)
 				Y_[i](j) = omega_gas(j);
 
-		for (int i = 0; i < np_; i++)
+		for (unsigned int i = 0; i < np_; i++)
 			for (unsigned int j = 0; j < surf_nc_; j++)
 				Z_[i](j) = Z0(j);
 
-		for (int i = 0; i < np_; i++)
+		for (unsigned int i = 0; i < np_; i++)
 			for (unsigned int j = 0; j < surf_np_; j++)
 			{
 				Gamma_[i](j) = Gamma0(j);
@@ -491,7 +491,7 @@ namespace CVI
 
 	void Reactor1D::Properties()
 	{
-		for (int i = 0; i < np_; i++)
+		for (unsigned int i = 0; i < np_; i++)
 		{
 			// Thermodynamics
 			{
@@ -728,7 +728,7 @@ namespace CVI
 	void Reactor1D::SubEquations_Porosity()
 	{
 		// Internal points
-		for (int i = 0; i < np_; i++)
+		for (unsigned int i = 0; i < np_; i++)
 		{
 			const double coefficient = 1000.;
 			const double smoothing_coefficient = 0.50*(std::tanh(coefficient*(epsilon_(i)-porousMedium_.epsilon_threshold()))+1.);
@@ -876,7 +876,7 @@ namespace CVI
 	void Reactor1D::Recover_Unknowns(const double* y)
 	{
 		unsigned int count = 0;
-		for (int i = 0; i < np_; i++)
+		for (unsigned int i = 0; i < np_; i++)
 		{
 			// Species
 			for (unsigned int j = 0; j < nc_; j++)
@@ -904,7 +904,7 @@ namespace CVI
 	void Reactor1D::Recover_Residuals(double* dy)
 	{
 		unsigned int count = 0;
-		for (int i = 0; i < np_; i++)
+		for (unsigned int i = 0; i < np_; i++)
 		{
 			// Species
 			for (unsigned int j = 0; j < nc_; j++)
@@ -936,7 +936,7 @@ namespace CVI
 	void Reactor1D::UnknownsVector(double* v)
 	{
 		unsigned int count = 0;
-		for (int i = 0; i < np_; i++)
+		for (unsigned int i = 0; i < np_; i++)
 		{
 			for (unsigned int j = 0; j < nc_; j++)
 				v[count++] = Y_[i](j);
@@ -954,7 +954,7 @@ namespace CVI
 	void Reactor1D::CorrectedUnknownsVector(double* v)
 	{
 		unsigned int count = 0;
-		for (int i = 0; i < np_; i++)
+		for (unsigned int i = 0; i < np_; i++)
 		{
 			for (unsigned int j = 0; j < nc_; j++)
 				Y_[i](j) = v[count++];
@@ -979,7 +979,7 @@ namespace CVI
 		const double zero = 0.;
 
 		unsigned int count = 0;
-		for (int i = 0; i < np_; i++)
+		for (unsigned int i = 0; i < np_; i++)
 		{
 			for (unsigned int j = 0; j < nc_; j++)
 				v[count++] = zero;
@@ -999,7 +999,7 @@ namespace CVI
 		const double one = 1.;
 
 		unsigned int count = 0;
-		for (int i = 0; i < np_; i++)
+		for (unsigned int i = 0; i < np_; i++)
 		{
 			for (unsigned int j = 0; j < nc_; j++)
 				v[count++] = one;
@@ -1159,7 +1159,7 @@ namespace CVI
 		std::cout << "--------------------------------------------------------------------------" << std::endl;
 
 		// Loop
-		unsigned int number_intervals = time_total_ / dae_time_interval_;
+		unsigned int number_intervals = static_cast<unsigned int>(time_total_ / dae_time_interval_);
 		for (unsigned int k = 1; k <= number_intervals; k++)
 		{
 			const double t0 = (k - 1)*dae_time_interval_;
@@ -1230,7 +1230,7 @@ namespace CVI
 			fOutput << std::endl;
 		}
 
-		for (int i = 0; i < np_; i++)
+		for (unsigned int i = 0; i < np_; i++)
 		{
 			OpenSMOKE::OpenSMOKEVectorDouble yy(nc_);
 			OpenSMOKE::OpenSMOKEVectorDouble xx(nc_);
@@ -1309,7 +1309,7 @@ namespace CVI
 			fOutput << std::endl;
 		}
 
-		for (int i = 0; i < np_; i++)
+		for (unsigned int i = 0; i < np_; i++)
 		{
 			// Calculate the diffusion coefficients
 			{
@@ -1367,7 +1367,7 @@ namespace CVI
 			fOutput << std::endl;
 		}
 
-		for (int i = 0; i < np_; i++)
+		for (unsigned int i = 0; i < np_; i++)
 		{
 			// Formation rates
 			{
@@ -1449,7 +1449,7 @@ namespace CVI
 			fOutput << std::endl;
 		}
 
-		for (int i = 0; i < np_; i++)
+		for (unsigned int i = 0; i < np_; i++)
 		{
 			// Calculate the reaction and formation rates of heterogeneous reactions
 			{
@@ -1548,7 +1548,7 @@ namespace CVI
 			fOutput << std::endl;
 		}
 
-		for (int i = 0; i < np_; i++)
+		for (unsigned int i = 0; i < np_; i++)
 		{
 			// Calculate the reaction and formation rates of heterogeneous reactions
 			{
@@ -1638,7 +1638,7 @@ namespace CVI
 
 			for (unsigned int i=0;i<list_points.size();i++)
 			{
-				const unsigned int point = list_points[i];
+				const unsigned int point = static_cast<unsigned int>(list_points[i]);
 
 				fROPA << std::endl;
 				fROPA << "-------------------------------------------------------------------" << std::endl;
@@ -1692,7 +1692,7 @@ namespace CVI
 
 				if (detailed_heterogeneous_kinetics_ == true)
 				{
-					for (int i = 0; i < surf_np_; i++)
+					for (unsigned int i = 0; i < surf_np_; i++)
 						std::cout << std::left << std::setw(16) << "Gamma[kmol/m2]";
 					std::cout << std::left << std::setw(14) << "Error";
 				}
@@ -1715,11 +1715,11 @@ namespace CVI
 
 			if (detailed_heterogeneous_kinetics_ == true)
 			{
-				for (int i = 0; i < surf_np_; i++)
+				for (unsigned int i = 0; i < surf_np_; i++)
 					std::cout << std::left << std::setw(16) << std::scientific << Gamma_[0](i);
 
 				Eigen::VectorXd error_z_sum(np_);
-				for (int i = 0; i < np_; i++)
+				for (unsigned int i = 0; i < np_; i++)
 					error_z_sum(i) = std::fabs(Z_[i].sum() - 1.);
 
 				std::cout << std::left << std::setw(14) << std::setprecision(2) << std::scientific << error_z_sum.maxCoeff();
